requirements.txt
attrs==21.2.0
isoweek==1.3.3
python-dateutil==2.8.1
sqlalchemy==1.4.15
pyaml-env==1.1.0